rename raw exec directory; result of rebasing to f-driver-plugins branch
