+++ plugins/drivers/raw-exec/Makefile
@@ -1,3 +1,7 @@
+deps:
+	go get -u github.com/golang/protobuf/proto
+	go get -u github.com/golang/protobuf/protoc-gen-go
+
 generate:
 	protoc -I proto/ proto/raw-exec.proto --go_out=plugins=grpc:proto/
