+++ plugins/drivers/raw-exec/main.go
@@ -7,8 +7,12 @@ import (
 
 	"github.com/hashicorp/go-plugin"
 
-	"github.com/hashicorp/nomad/plugins/drivers/raw-exec/proto"
-	"github.com/hashicorp/nomad/plugins/drivers/raw-exec/shared"
+	"strings"
+
+	"github.com/golang/protobuf/jsonpb"
+	_struct "github.com/golang/protobuf/ptypes/struct"
+	"github.com/hashicorp/nomad/plugins/drivers/raw_exec/proto"
+	"github.com/hashicorp/nomad/plugins/drivers/raw_exec/shared"
 )
 
 func main() {
@@ -37,13 +41,51 @@ func main() {
 	}
 
 	rawExec := raw.(shared.RawExec)
-	execCtx := &proto.ExecContext{}
-	taskInfo := &proto.TaskInfo{}
 
-	result, err := rawExec.Start(execCtx, taskInfo)
+	currentDir, err := os.Getwd() // TODO
+	if err != nil {
+		panic(fmt.Sprintf("encoungered error when getting current dir: %s", err.Error()))
+	}
+
+	execCtx := &proto.ExecContext{
+		TaskDir: &proto.TaskDir{
+			Directory: currentDir,
+			LogDir:    currentDir,
+			LogLevel:  "DEBUG",
+		},
+		TaskEnv: &proto.TaskEnv{},
+	}
+	jsonConfig := `{
+                    "Command":"echo",
+                    "Args":["the", "quick", "brown", "fox", "jumped"]
+                   }`
+	unMarshaller := jsonpb.Unmarshaler{AllowUnknownFields: false}
+
+	reader := strings.NewReader(jsonConfig)
+	structConfig := &_struct.Struct{}
+	err = unMarshaller.Unmarshal(reader, structConfig)
+
+	if err != nil {
+		fmt.Println("Error unmarshalling json into protobuf Struct:%v", err)
+		os.Exit(-1)
+	}
+	taskInfo := &proto.TaskInfo{
+		Resources: &proto.Resources{
+			Cpu:      250,
+			MemoryMb: 256,
+			DiskMb:   20,
+		},
+		LogConfig: &proto.LogConfig{
+			MaxFiles:      10,
+			MaxFileSizeMb: 10,
+		},
+		Config: structConfig,
+	}
+
+	result, err := rawExec.NewStart(execCtx, taskInfo)
 	if err != nil {
 		fmt.Printf("Encountered errors: %s \n", err.Error())
 	}
 
-	fmt.Printf(" Found task id: %s \n", result.TaskId)
+	fmt.Printf(": %s \n", result)
 }
