+++ plugins/drivers/raw-exec/proto/raw-exec.proto
@@ -1,5 +1,6 @@
 syntax = "proto3";
 package proto;
+import "google/protobuf/struct.proto";
 
 message Port {
   string label = 1;
@@ -31,7 +32,9 @@ message LogConfig {
 message TaskInfo {
   Resources resources = 1;
   LogConfig log_config = 2;
-  Struct config = 3;
+  //Struct config = 3;
+  // example: https://gist.github.com/jsmouret/2bc876e8def6c63410556350eca3e43d
+  google.protobuf.Struct config = 3;
 }
 
 message TaskDir {
@@ -41,6 +44,7 @@ message TaskDir {
   string local_dir = 4;
   string log_dir = 5;
   string secrets_dir = 6;
+  string log_level = 7;
 }
 
 message TaskEnv {
@@ -72,60 +76,5 @@ message StartResponse {
 message Empty {}
 
 service RawExec {
-    rpc Start(StartRequest) returns (StartResponse);
-}
-
-// TODO this needs to be resolved and part of a different proto file.
-// `Struct` represents a structured data value, consisting of fields
-// which map to dynamically typed values. In some languages, `Struct`
-// might be supported by a native representation. For example, in
-// scripting languages like JS a struct is represented as an
-// object. The details of that representation are described together
-// with the proto support for the language.
-//
-// The JSON representation for `Struct` is JSON object.
-message Struct {
-  // Unordered map of dynamically typed values.
-  map<string, Value> fields = 1;
-}
-
-// `Value` represents a dynamically typed value which can be either
-// null, a number, a string, a boolean, a recursive struct value, or a
-// list of values. A producer of value is expected to set one of that
-// variants, absence of any variant indicates an error.
-//
-// The JSON representation for `Value` is JSON value.
-message Value {
-  // The kind of value.
-  oneof kind {
-    // Represents a null value.
-    NullValue null_value = 1;
-    // Represents a double value.
-    double number_value = 2;
-    // Represents a string value.
-    string string_value = 3;
-    // Represents a boolean value.
-    bool bool_value = 4;
-    // Represents a structured value.
-    Struct struct_value = 5;
-    // Represents a repeated `Value`.
-    ListValue list_value = 6;
-  }
-}
-
-// `NullValue` is a singleton enumeration to represent the null value for the
-// `Value` type union.
-//
-//  The JSON representation for `NullValue` is JSON `null`.
-enum NullValue {
-  // Null value.
-  NULL_VALUE = 0;
-}
-
-// `ListValue` is a wrapper around a repeated field of values.
-//
-// The JSON representation for `ListValue` is JSON array.
-message ListValue {
-  // Repeated field of dynamically typed values.
-  repeated Value values = 1;
+    rpc NewStart(StartRequest) returns (StartResponse);
 }
