--- plugins/drivers/raw_exec_driver/proto/raw-exec.proto
@@ -1,80 +1,0 @@
-syntax = "proto3";
-package proto;
-import "google/protobuf/struct.proto";
-
-message Port {
-  string label = 1;
-  uint32 value = 2;
-}
-
-message Networks {
-  string device = 1;
-  string cidr = 2;
-  string ip = 3;
-  uint64 mbits = 4;
-  Port reserved_ports = 5;
-  Port dynamic_ports = 6;
-}
-
-message Resources {
-  uint64 cpu = 1;
-  uint64 memory_mb = 2;
-  uint64 disk_mb = 3;
-  uint64 iops = 4;
-  Networks networks = 5;
-}
-
-message LogConfig {
-  uint32 max_files = 1;
-  uint64 max_file_size_mb = 2;
-}
-
-message TaskInfo {
-  Resources resources = 1;
-  LogConfig log_config = 2;
-  //Struct config = 3;
-  // example: https://gist.github.com/jsmouret/2bc876e8def6c63410556350eca3e43d
-  google.protobuf.Struct config = 3;
-}
-
-message TaskDir {
-  string directory = 1;
-  string shared_alloc_dir = 2;
-  string shared_task_dir = 3;
-  string local_dir = 4;
-  string log_dir = 5;
-  string secrets_dir = 6;
-  string log_level = 7;
-}
-
-message TaskEnv {
-  map<string, string> node_attrs = 1;
-  map<string, string> env_map = 2;
-}
-
-message ExecContext {
-  TaskDir task_dir = 1;
-  TaskEnv task_env = 2;
-}
-
-message StartRequest {
-  ExecContext exec_context = 1;
-  TaskInfo task_info = 2;
-}
-
-message DriverNetwork {
-  map<string, uint32> port_map = 1;
-  string ip = 2;
-  bool auto_advertise = 3;
-}
-
-message StartResponse {
-  string task_id = 1;
-  DriverNetwork driver_network = 2;
-}
-
-message Empty {}
-
-service RawExec {
-    rpc NewStart(StartRequest) returns (StartResponse);
-}
